- [x] Xml filename changed, updated in database and published in registry when publisher id is changed. - [x] Xml updated, database and registry record updated when recipient country code is changed. - [x] Twitter auto publish migration - [x] Queue implementation for changing publisher id and publishing workflow - [x] Send email to notify user when publisher id and files have been changed

+++ app/Http/Controllers/Lite/PublishedFiles/PublishedFilesController.php
@@ -5,6 +5,7 @@ use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\RedirectResponse;
 use App\Http\Controllers\Lite\LiteController;
 use App\Lite\Services\PublishedFiles\PublishedFilesService;
+use Illuminate\Support\Facades\Session;
 
 /**
  * Class PublishedFilesController
@@ -16,6 +17,7 @@ class PublishedFilesController extends LiteController
      * @var PublishedFilesService
      */
     protected $publishedFilesService;
+
 
     /**
      * PublishedFilesController constructor.
@@ -54,6 +56,12 @@ class PublishedFilesController extends LiteController
         }
 
         $this->authorize('delete_activity', $file);
+
+        if (Session::has('publisherIdChange') || isPublisherIdBeingChanged()) {
+            $response = ['type' => 'success', 'code' => ['message', ['message' => trans('success.publisher_id_changing')]]];
+
+            return redirect()->to('publishing-settings')->withResponse($response);
+        }
 
         if (!$this->publishedFilesService->delete($id)) {
             $response = ['type' => 'danger', 'code' => ['message', ['message' => trans('lite/global.delete_unsuccessful', ['type' => 'File'])]]];
@@ -96,3 +104,4 @@ class PublishedFilesController extends LiteController
         return redirect()->back()->withResponse(['type' => 'success', 'code' => ['message', ['message' => trans('lite/global.bulk_publish_successful')]]]);
     }
 }
+
